url/actionOpts can be functions now
Instead of specifying the value for the resource configuration directly, you can now specify a function that will be called at dispatch time to determine the final {url, headers,method,…).

This allows for URLs that are not known at build time but at run time, such as resources that are discovered from fetching other resources.